Refactor: BudgetContainer 수정
